# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "5/3/1 Training App",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workouts": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "List workouts",
        "description": "Get workouts for the current user with optional filters.",
        "operationId": "list_workouts_api_v1_workouts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "program_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by program ID",
              "title": "Program Id"
            },
            "description": "Filter by program ID"
          },
          {
            "name": "workout_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (scheduled, completed, skipped)",
              "title": "Workout Status"
            },
            "description": "Filter by status (scheduled, completed, skipped)"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by date >= start_date",
              "title": "Start Date"
            },
            "description": "Filter by date >= start_date"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by date <= end_date",
              "title": "End Date"
            },
            "description": "Filter by date <= end_date"
          },
          {
            "name": "main_lifts",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by main lifts (press, deadlift, bench_press, squat)",
              "title": "Main Lifts"
            },
            "description": "Filter by main lifts (press, deadlift, bench_press, squat)"
          },
          {
            "name": "cycle_number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by cycle number",
              "title": "Cycle Number"
            },
            "description": "Filter by cycle number"
          },
          {
            "name": "week_number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 4,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by week number (1-4)",
              "title": "Week Number"
            },
            "description": "Filter by week number (1-4)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutResponse"
                  },
                  "title": "Response List Workouts Api V1 Workouts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workouts/missed": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get missed workouts",
        "description": "Get all workouts that are past their scheduled date but not completed.",
        "operationId": "get_missed_workouts_api_v1_workouts_missed_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissedWorkoutsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/workouts/{workout_id}": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get workout details",
        "description": "Get detailed workout information including all prescribed sets.",
        "operationId": "get_workout_api_v1_workouts__workout_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workouts/{workout_id}/complete": {
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Complete workout",
        "description": "Log all sets, mark workout as completed, and get performance analysis.",
        "operationId": "complete_workout_api_v1_workouts__workout_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutCompletionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workouts/{workout_id}/skip": {
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Skip workout",
        "description": "Mark a workout as intentionally skipped.",
        "operationId": "skip_workout_api_v1_workouts__workout_id__skip_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workouts/{workout_id}/handle-missed": {
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Handle missed workout",
        "description": "Skip or reschedule a missed workout.",
        "operationId": "handle_missed_workout_api_v1_workouts__workout_id__handle_missed_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandleMissedWorkoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandleMissedWorkoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HandleMissedWorkoutRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action",
            "description": "Action to take: 'skip' or 'reschedule'"
          },
          "reschedule_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reschedule Date",
            "description": "Date to reschedule to (required if action is 'reschedule')"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "HandleMissedWorkoutRequest",
        "description": "Request to handle a missed workout."
      },
      "HandleMissedWorkoutResponse": {
        "properties": {
          "workout": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkoutResponse"
              }
            ],
            "description": "The updated workout"
          },
          "action_taken": {
            "type": "string",
            "title": "Action Taken",
            "description": "Action that was taken"
          },
          "rescheduled_count": {
            "type": "integer",
            "title": "Rescheduled Count",
            "description": "Number of workouts rescheduled (including cascaded)",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "workout",
          "action_taken"
        ],
        "title": "HandleMissedWorkoutResponse",
        "description": "Response after handling a missed workout."
      },
...
      "MissedWorkoutInfo": {
        "properties": {
          "workout": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkoutResponse"
              }
            ],
            "description": "The missed workout"
          },
          "days_overdue": {
            "type": "integer",
            "title": "Days Overdue",
            "description": "Days since scheduled date"
          },
          "can_reschedule": {
            "type": "boolean",
            "title": "Can Reschedule",
            "description": "Whether rescheduling is possible"
          }
        },
        "type": "object",
        "required": [
          "workout",
          "days_overdue",
          "can_reschedule"
        ],
        "title": "MissedWorkoutInfo",
        "description": "Information about a missed workout."
      },
...
      "MissedWorkoutsResponse": {
        "properties": {
          "missed_workouts": {
            "items": {
              "$ref": "#/components/schemas/MissedWorkoutInfo"
            },
            "type": "array",
            "title": "Missed Workouts",
            "description": "List of missed workouts"
          },
          "user_preference": {
            "type": "string",
            "title": "User Preference",
            "description": "User's missed workout preference (skip/reschedule/ask)"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of missed workouts"
          }
        },
        "type": "object",
        "required": [
          "missed_workouts",
          "user_preference",
          "count"
        ],
        "title": "MissedWorkoutsResponse",
        "description": "Response containing all missed workouts."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkoutCompleteRequest": {
        "properties": {
          "sets": {
            "items": {
              "$ref": "#/components/schemas/SetLogRequest"
            },
            "type": "array",
            "minItems": 1,
            "title": "Sets",
            "description": "All sets performed"
          },
          "workout_notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Notes"
          },
          "completed_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Date",
            "description": "When workout was completed (defaults to now)"
          }
        },
        "type": "object",
        "required": [
          "sets"
        ],
        "title": "WorkoutCompleteRequest",
        "description": "Request to complete a workout with all logged sets."
      },
      "WorkoutCompletionResponse": {
        "properties": {
          "workout": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkoutResponse"
              }
            ],
            "description": "The completed workout"
          },
          "analysis": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkoutAnalysis"
              }
            ],
            "description": "Analysis of workout performance"
          }
        },
        "type": "object",
        "required": [
          "workout",
          "analysis"
        ],
        "title": "WorkoutCompletionResponse",
        "description": "Response for workout completion including analysis."
      },
      "WorkoutDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "program_id": {
            "type": "string",
            "title": "Program Id"
          },
          "scheduled_date": {
            "type": "string",
            "format": "date",
            "title": "Scheduled Date"
          },
          "completed_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Date"
          },
          "cycle_number": {
            "type": "integer",
            "title": "Cycle Number"
          },
          "week_number": {
            "type": "integer",
            "title": "Week Number"
          },
          "week_type": {
            "type": "string",
            "title": "Week Type"
          },
          "main_lifts": {
            "items": {
              "$ref": "#/components/schemas/WorkoutMainLiftResponse"
            },
            "type": "array",
            "title": "Main Lifts"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "sets_by_lift": {
            "additionalProperties": {
              "$ref": "#/components/schemas/WorkoutSetsForLift"
            },
            "type": "object",
            "title": "Sets By Lift"
          },
          "accessory_sets": {
            "items": {
              "$ref": "#/components/schemas/WorkoutSetResponse"
            },
            "type": "array",
            "title": "Accessory Sets"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "program_id",
          "scheduled_date",
          "completed_date",
          "cycle_number",
          "week_number",
          "week_type",
          "main_lifts",
          "status",
          "sets_by_lift",
          "accessory_sets",
          "notes",
          "created_at"
        ],
        "title": "WorkoutDetailResponse",
        "description": "Detailed workout with prescribed sets organized by lift."
      },
...
      "WorkoutMainLiftResponse": {
        "properties": {
          "lift_type": {
            "type": "string",
            "title": "Lift Type"
          },
          "lift_order": {
            "type": "integer",
            "title": "Lift Order"
          },
          "current_training_max": {
            "type": "number",
            "title": "Current Training Max"
          }
        },
        "type": "object",
        "required": [
          "lift_type",
          "lift_order",
          "current_training_max"
        ],
        "title": "WorkoutMainLiftResponse",
        "description": "Response model for a main lift in a workout."
      },
      "WorkoutResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "program_id": {
            "type": "string",
            "title": "Program Id"
          },
          "scheduled_date": {
            "type": "string",
            "format": "date",
            "title": "Scheduled Date"
          },
          "completed_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Date"
          },
          "cycle_number": {
            "type": "integer",
            "title": "Cycle Number"
          },
          "week_number": {
            "type": "integer",
            "title": "Week Number"
          },
          "week_type": {
            "type": "string",
            "title": "Week Type"
          },
          "main_lifts": {
            "items": {
              "$ref": "#/components/schemas/WorkoutMainLiftResponse"
            },
            "type": "array",
            "title": "Main Lifts"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "program_id",
          "scheduled_date",
          "completed_date",
          "cycle_number",
          "week_number",
          "week_type",
          "main_lifts",
          "status",
          "notes",
          "created_at"
        ],
        "title": "WorkoutResponse",
        "description": "Basic workout information."
      },
      "WorkoutSetResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "set_type": {
            "type": "string",
            "title": "Set Type"
          },
          "set_number": {
            "type": "integer",
            "title": "Set Number"
          },
          "prescribed_reps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prescribed Reps"
          },
          "prescribed_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prescribed Weight"
          },
          "percentage_of_tm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentage Of Tm"
          },
          "actual_reps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Reps"
          },
          "actual_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Weight"
          },
          "is_target_met": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Target Met"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "exercise_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exercise Id"
          },
          "circuit_group": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Circuit Group"
          }
        },
        "type": "object",
        "required": [
          "set_type",
          "set_number",
          "prescribed_reps",
          "prescribed_weight",
          "percentage_of_tm"
        ],
        "title": "WorkoutSetResponse",
        "description": "Response model for a workout set (prescribed or logged)."
      },
      "WorkoutSetsForLift": {
        "properties": {
          "warmup_sets": {
            "items": {
              "$ref": "#/components/schemas/WorkoutSetResponse"
            },
            "type": "array",
            "title": "Warmup Sets"
          },
          "main_sets": {
            "items": {
              "$ref": "#/components/schemas/WorkoutSetResponse"
            },
            "type": "array",
            "title": "Main Sets"
          }
        },
        "type": "object",
        "required": [
          "warmup_sets",
          "main_sets"
        ],
        "title": "WorkoutSetsForLift",
        "description": "Sets organized by type for a specific lift."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}